refactor: FlexBox 컴포넌트 type 파일 이름 변경
