+++ frontend/src/components/@shared/FlexBox/type.ts
